// cypress/data_objects/account-dto.ts
export interface AccountDto {
    title: string;
    name: string;
    email: string;
    password: string;
    day: string;
    month: string;
    year: string;
    first_name: string;
    last_name: string;
    company: string;
    address: string;
    address2: string;
    country: string;
    state: string;
    city: string;
    zipcode: string;
    mobile_number: string;
    newsletter?: boolean;
    optin?: boolean;
}